ml/serializers.py
from rest_framework import serializers
from .models import MLModel


class MLModelSerializer(serializers.ModelSerializer):
    class Meta:
        model = MLModel
        fields = '__all__'
        read_only_fields = ('id', 'created_at', 'status', 'accuracy')
