requirements.txt
django
djangorestframework
django-cors-headers
pandas
numpy
tensorflow
scikit-learn
channels
channels_redis
python-decouple
celery
daphne
redis
gunicorn